requirements.txt
aiogram==3.4.1
SQLAlchemy==2.0.30
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pandas==2.2.2
openpyxl==3.1.2
